+++ lib/grinder_sdk.dart
@@ -12,6 +12,7 @@ import 'package:which/which.dart';
 
 import 'grinder.dart';
 import 'src/run.dart' as run_lib;
+import 'src/run_utils.dart';
 
 bool _sdkOnPath;
 
@@ -42,10 +43,12 @@ class Dart {
   /// Run a dart [script] using [run_lib.run].
   ///
   /// Returns the stdout.
-  static String run(String script,
-      {List<String> arguments : const [], bool quiet: false,
-       String packageRoot, String workingDirectory, int vmNewGenHeapMB,
+  static String run(String script, {List<String> arguments : const [],
+      bool quiet: false, String packageRoot, RunOptions runOptions,
+      @Deprecated('Use RunOptions.workingDirectory instead.')
+      String workingDirectory, int vmNewGenHeapMB,
        int vmOldGenHeapMB, List<String> vmArgs : const []}) {
+    runOptions = mergeWorkingDirectory(workingDirectory, runOptions);
     List<String> args = [];
 
     args.addAll(vmArgs);
@@ -66,7 +69,7 @@ class Dart {
     args.addAll(arguments);
 
     return run_lib.run(_sdkBin('dart'), arguments: args, quiet: quiet,
-        workingDirectory: workingDirectory);
+        runOptions: runOptions);
   }
 
   static String version({bool quiet: false}) {
@@ -96,12 +99,15 @@ class Pub {
    * even if the pubspec.lock file is up-to-date with respect to the
    * pubspec.yaml file.
    */
-  static void get({bool force: false, String workingDirectory}) {
+  static void get({bool force: false, RunOptions runOptions,
+      @Deprecated('Use RunOptions.workingDirectory instead.')
+      String workingDirectory}) {
+    runOptions = mergeWorkingDirectory(workingDirectory, runOptions);
     FileSet pubspec = new FileSet.fromFile(new File('pubspec.yaml'));
     FileSet publock = new FileSet.fromFile(new File('pubspec.lock'));
 
     if (force || !publock.upToDate(pubspec)) {
-      _run('get', workingDirectory: workingDirectory);
+      _run('get', runOptions: runOptions);
     }
   }
 
@@ -110,13 +116,16 @@ class Pub {
    * even if the pubspec.lock file is up-to-date with respect to the
    * pubspec.yaml file.
    */
-  static Future getAsync({bool force: false, String workingDirectory}) {
+  static Future getAsync({bool force: false, RunOptions runOptions,
+      @Deprecated('Use RunOptions.workingDirectory instead.')
+      String workingDirectory}) {
+    runOptions = mergeWorkingDirectory(workingDirectory, runOptions);
     FileSet pubspec = new FileSet.fromFile(new File('pubspec.yaml'));
     FileSet publock = new FileSet.fromFile(new File('pubspec.lock'));
 
     if (force || !publock.upToDate(pubspec)) {
       return run_lib.runAsync(_sdkBin('pub'), arguments: ['get'],
-          workingDirectory: workingDirectory).then((_) => null);
+      runOptions: runOptions).then((_) => null);
     }
 
     return new Future.value();
@@ -125,31 +134,43 @@ class Pub {
   /**
    * Run `pub upgrade` on the current project.
    */
-  static void upgrade({String workingDirectory}) {
-    _run('upgrade', workingDirectory: workingDirectory);
+  static void upgrade({RunOptions runOptions,
+      @Deprecated('Use RunOptions.workingDirectory instead.')
+      String workingDirectory}) {
+    runOptions = mergeWorkingDirectory(workingDirectory, runOptions);
+    _run('upgrade', runOptions: runOptions);
   }
 
   /**
    * Run `pub upgrade` on the current project.
    */
-  static Future upgradeAsync({String workingDirectory}) {
+  static Future upgradeAsync({RunOptions runOptions,
+      @Deprecated('Use RunOptions.workingDirectory instead.')
+      String workingDirectory}) {
+    runOptions = mergeWorkingDirectory(workingDirectory, runOptions);
     return run_lib.runAsync(_sdkBin('pub'), arguments: ['upgrade'],
-        workingDirectory: workingDirectory).then((_) => null);
+        runOptions: runOptions).then((_) => null);
   }
 
   /**
    * Run `pub downgrade` on the current project.
    */
-  static void downgrade({String workingDirectory}) {
-    _run('downgrade', workingDirectory: workingDirectory);
+  static void downgrade({RunOptions runOptions,
+      @Deprecated('Use RunOptions.workingDirectory instead.')
+      String workingDirectory}) {
+    runOptions = mergeWorkingDirectory(workingDirectory, runOptions);
+    _run('downgrade', runOptions: runOptions);
   }
 
   /**
    * Run `pub downgrade` on the current project.
    */
-  static Future downgradeAsync({String workingDirectory}) {
+  static Future downgradeAsync({RunOptions runOptions,
+      @Deprecated('Use RunOptions.workingDirectory instead.')
+      String workingDirectory}) {
+    runOptions = mergeWorkingDirectory(workingDirectory, runOptions);
     return run_lib.runAsync(_sdkBin('pub'), arguments: ['downgrade'],
-        workingDirectory: workingDirectory).then((_) => null);
+        runOptions: runOptions).then((_) => null);
   }
 
   /**
@@ -160,15 +181,18 @@ class Pub {
   static void build({
       String mode,
       List<String> directories,
-      String workingDirectory,
-      String outputDirectory}) {
+      RunOptions runOptions,
+      String outputDirectory,
+      @Deprecated('Use RunOptions.workingDirectory instead.')
+      String workingDirectory}) {
+    runOptions = mergeWorkingDirectory(workingDirectory, runOptions);
     List args = ['build'];
     if (mode != null) args.add('--mode=${mode}');
     if (outputDirectory != null) args.add('--output=${outputDirectory}');
     if (directories != null && directories.isNotEmpty) args.addAll(directories);
 
     run_lib.run(_sdkBin('pub'), arguments: args,
-        workingDirectory: workingDirectory);
+    runOptions: runOptions);
   }
 
   /**
@@ -179,27 +203,33 @@ class Pub {
   static Future buildAsync({
       String mode,
       List<String> directories,
-      String workingDirectory,
-      String outputDirectory}) {
+      RunOptions runOptions,
+      String outputDirectory,
+      @Deprecated('Use RunOptions.workingDirectory instead.')
+      String workingDirectory}) {
+    runOptions = mergeWorkingDirectory(workingDirectory, runOptions);
     List args = ['build'];
     if (mode != null) args.add('--mode=${mode}');
     if (outputDirectory != null) args.add('--output=${outputDirectory}');
     if (directories != null && directories.isNotEmpty) args.addAll(directories);
 
     return run_lib.runAsync(_sdkBin('pub'), arguments: args,
-        workingDirectory: workingDirectory).then((_) => null);
+        runOptions: runOptions).then((_) => null);
   }
 
   /// Run `pub run` on the given [package] and [script].
   ///
   /// If [script] is null it defaults to the same value as [package].
   static String run(String package,
-      {List<String> arguments, String workingDirectory, String script}) {
+      {List<String> arguments, RunOptions runOptions, String script,
+      @Deprecated('Use RunOptions.workingDirectory instead.')
+      String workingDirectory}) {
+    runOptions = mergeWorkingDirectory(workingDirectory, runOptions);
     var scriptArg = script == null ? package : '$package:$script';
     List args = ['run', scriptArg];
     if (arguments != null) args.addAll(arguments);
     return run_lib.run(_sdkBin('pub'), arguments: args,
-        workingDirectory: workingDirectory);
+    runOptions: runOptions);
   }
 
   static String version({bool quiet: false}) => _AppVersion.parse(
@@ -207,9 +237,9 @@ class Pub {
 
   static PubGlobal get global => _global;
 
-  static String _run(String command, {bool quiet: false, String workingDirectory}) {
+  static String _run(String command, {bool quiet: false, RunOptions runOptions}) {
     return run_lib.run(_sdkBin('pub'), quiet: quiet, arguments: [command],
-        workingDirectory: workingDirectory);
+        runOptions: runOptions);
   }
 }
 
@@ -327,12 +357,15 @@ class PubGlobal {
 
   /// Run the given installed Dart application.
   String run(String package,
-      {List<String> arguments, String workingDirectory, String script}) {
+      {List<String> arguments, RunOptions runOptions, String script,
+      @Deprecated('Use RunOptions.workingDirectory instead.')
+      String workingDirectory}) {
+    runOptions = mergeWorkingDirectory(workingDirectory, runOptions);
     var scriptArg = script == null ? package : '$package:$script';
     List args = ['global', 'run', scriptArg];
     if (arguments != null) args.addAll(arguments);
     return run_lib.run(_sdkBin('pub'), arguments: args,
-        workingDirectory: workingDirectory);
+        runOptions: runOptions);
   }
 
   /// Return the list of installed applications.
@@ -397,7 +430,9 @@ abstract class PubApp {
   /// If [script] is provided, the sub-script will be run. So
   /// `new PubApp.global('grinder').run(script: 'init');` will run
   /// `grinder:init`.
-  String run(List<String> arguments, {String script, String workingDirectory});
+  String run(List<String> arguments, {String script, RunOptions runOptions,
+      @Deprecated('Use RunOptions.workingDirectory instead.')
+      String workingDirectory});
 
   String toString() => packageName;
 }
@@ -446,13 +481,16 @@ class _PubGlobalApp extends PubApp {
   void activate({bool force: false}) =>
       Pub.global.activate(packageName, force: force);
 
-  String run(List<String> arguments, {String script, String workingDirectory}) {
+  String run(List<String> arguments, {String script, RunOptions runOptions,
+      @Deprecated('Use RunOptions.workingDirectory instead.')
+      String workingDirectory}) {
+    runOptions = mergeWorkingDirectory(workingDirectory, runOptions);
     activate();
 
     return Pub.global.run(packageName,
         script: script,
         arguments: arguments,
-        workingDirectory: workingDirectory);
+        runOptions: runOptions);
   }
 }
 
@@ -466,10 +504,13 @@ class _PubLocalApp extends PubApp {
 
   void activate({bool force: false}) { }
 
-  String run(List<String> arguments, {String script, String workingDirectory}) {
+  String run(List<String> arguments, {String script, RunOptions runOptions,
+      @Deprecated('Use RunOptions.workingDirectory instead.')
+      String workingDirectory}) {
+    runOptions = mergeWorkingDirectory(workingDirectory, runOptions);
     return Pub.run(packageName,
         script: script,
         arguments: arguments,
-        workingDirectory: workingDirectory);
+        runOptions: runOptions);
   }
 }